+++ R/utils_model_parameters.R
@@ -159,7 +159,7 @@
   if ("ci_digits" %in% names(dot.arguments)) {
     attr(params, "ci_digits") <- dot.arguments[["ci_digits"]]
   } else {
-    attr(params, "ci_digits") <- 2
+    attr(params, "ci_digits") <- NULL
   }
 
   if ("p_digits" %in% names(dot.arguments)) {
@@ -213,16 +213,16 @@
 
   switch(tolower(ci_method),
     # abbreviations
-    "eti" = ,
-    "hdi" = ,
-    "si" = toupper(ci_method),
+    eti = ,
+    hdi = ,
+    si = toupper(ci_method),
     # named after people
-    "satterthwaite" = ,
-    "kenward" = ,
-    "wald" = insight::format_capitalize(ci_method),
+    satterthwaite = ,
+    kenward = ,
+    wald = insight::format_capitalize(ci_method),
     # special cases
-    "bci" = ,
-    "bcai" = "BCa",
+    bci = ,
+    bcai = "BCa",
     # no change otherwise
     ci_method
   )
@@ -236,12 +236,12 @@
     name <- attributes(s)$estName
     if (!is.null(name)) {
       coef_col <- switch(name,
-        "prob"       = "Probability",
-        "odds.ratio" = "Odds Ratio",
-        "emmean"     = "Marginal Means",
-        "rate"       = "Estimated Counts",
-        "ratio"      = "Ratio",
-        "Coefficient"
+        prob       = "Probability",
+        odds.ratio = "Odds Ratio",
+        emmean     = "Marginal Means",
+        rate       = "Estimated Counts",
+        ratio      = "Ratio",
+        Coefficient
       )
     }
   } else if (!is.null(info) && !info$family == "unknown") {
@@ -300,7 +300,17 @@
     return(params)
   }
 
-  columns <- grepl(pattern = "^(Coefficient|Mean|Median|MAP|Std_Coefficient|CI_|Std_CI)", colnames(params))
+  # pattern for marginaleffects objects
+  if (!is.null(attr(params, "coefficient_name"))) {
+    pattern <- sprintf(
+      "^(Coefficient|Mean|Median|MAP|Std_Coefficient|%s|CI_|Std_CI)",
+      attr(params, "coefficient_name")
+    )
+  } else {
+    pattern <- "^(Coefficient|Mean|Median|MAP|Std_Coefficient|CI_|Std_CI)"
+  }
+
+  columns <- grepl(pattern = pattern, colnames(params))
   if (any(columns)) {
     if (inherits(model, "mvord")) {
       rows <- params$Component != "correlation"
@@ -384,7 +394,7 @@
   if ("ci_digits" %in% names(dot.arguments)) {
     attr(params, "ci_digits") <- eval(dot.arguments[["ci_digits"]])
   } else {
-    attr(params, "ci_digits") <- 2
+    attr(params, "ci_digits") <- NULL
   }
 
   if ("p_digits" %in% names(dot.arguments)) {